Cleaned code and moved TokenClaims struct to prelude
