+++ src/lib.rs
@@ -63,5 +63,7 @@ mod builders;
 mod client;
 mod errors;
 pub mod groups;
+#[cfg(feature = "validate_jwt")]
 mod jwt_util;
 pub mod prelude;
+
